pyproject.toml
[tool.poetry]
name = "hr-space-project"
version = "0.1.0"
description = ""
authors = ["Pvl1307 <ivanov.pavel.1307@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2.15"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
ydb = "^3.8.0"
pillow = "^10.2.0"
djangorestframework = "^3.14.0"
celery = "^5.3.6"
redis = "^5.0.1"
django-celery-beat = "^2.5.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
